Initial commit
The pipeline is functional. It can be initiated by running either the NextFlow definition directly (for local testing) or the shell script for running on Viking.
This just loads the appropriate modules and Python virtual environments before running the pipeline.

The next step is to get the Slurm executor working so the segmentation is done on compute nodes.