Removed 'Infinity' parameter from JSON
Getting values like 'Infinity' correctly parsed as a float in Java will be tricky,
as it will be hard to differentiate between what's intentionally a string and what
is a numeric type. This could be implemented by looking for keywards such as `Infinity`
or `NaN`, but for now it's simpler to just leave these parameters out, especially considering these are the defaults.