TrackMate XML saving to disk
It appears that cause of the OOM memory issue was passing the entire XML contents as a string from Java to Python.
This commit saves the XML to disk so future processes will be needed to parse this efficiently. I'll initially try with the same Python parser as before, but if this still takes up too much memory then alternatives can be explored, such as xpath for the extracting of relevant parts to be processed in Python