feat: config file (#33)
* Can pass CellPhe features through config file

* Added cellpose parameters

* Added TrackMate options to config

* Added example iolight config

* Fixed json syntax error

* Fixed missing comma

* Escape JSON in quotation marks

* Pass the config file straight to Nextflow

* Removed 'Infinity' parameter from JSON

Getting values like 'Infinity' correctly parsed as a float in Java will be tricky,
as it will be hard to differentiate between what's intentionally a string and what
is a numeric type. This could be implemented by looking for keywards such as `Infinity`
or `NaN`, but for now it's simpler to just leave these parameters out, especially considering these are the defaults.

* Fixed bug with requesting cpus from Slurm

The `cpus` Nextflow option is being ignored by Slurm, so instead we explicitly request
the required number of cpus using Slurm specific options.
I'm not sure if this is a bug in Nextflow, or a difference in our Slurm configuration.

* Bumped up default resources

* Increased minimum cell size

Too small ROIs were identified which caused problems downstream when extracting features.

* Filter out small cells before filtering too short tracks

Previously tracks were filtered for being too short, _then_ cells were filtered
for being too small. This meant that some tracks were kept that once the small cells
were removed, weren't long enough and caused problems when creating the time series features.

The order has now been corrected and placed into a single QC step to make the filtering
logic more obvious.

Ideally the small ROIs would be filtered before the tracking. This is technically possible,
it would just require modifying the masks.

* Check for whether exception was OOM

Previously all errors were raised as OOM, when this wasn't necessarily the cause.
This means processes failing in an unrecoverable way, such as a syntax issue, won't be unncessarily re-attempted.

* Use config.json file that should be in the dataset folder

* Made templates dir for configs

* Copy config file over to Viking

* Always export outputs to network share

In case of pipeline failure, at least the intermediate outputs will be
available for debugging

* Fixed typo in feature names

Feature names MUST BE UPPER CASE!

* TrackMate XML saving to disk

It appears that cause of the OOM memory issue was passing the entire XML contents as a string from Java to Python.
This commit saves the XML to disk so future processes will be needed to parse this efficiently. I'll initially try with the same Python parser as before, but if this still takes up too much memory then alternatives can be explored, such as xpath for the extracting of relevant parts to be processed in Python

* Parse TrackMate XML

* Parsing XML into ROIs and CSV

* ROI filenames specified by Spot ID, Track ID and Frame ID

This will handle collisions between multiple cells having the same trackid in the same frame.