Don't update bioldata files modification time
If a Nextflow Pipeline uses cached outputs, such as a mask or CSV file,
it will publish this again to the output directory, updating the
modification time but the contents will be identical. Rclone by default
will update the remote copy with the new modification time. However,
this isn't that useful to us and in fact could be misleading if some
files have a later modification time than when they were actually
generated.