Replaced R CRAN module with custom library
We had a problem in that R doesn't have a neat package manager like pip or
virtual environment manager like venv, yet we needed multiple users to have access to
the exact same R packages when running the pipeline.

Initially I tried to setup a library of packages in the shared project dir, but I was unable to make this the default location as the `R_LIBS_USER` env var that determines where R should find packages
wasn't being read by R - it seemingly only reads it in if it's provided in a .Renviron
file, which resides in a user's home directory.
Rather than requiring every user of the pipeline to install the same R packages, we used
the R-CRAN EasyBuild module, which provides a whole host of packages.
However, while this solved the R problem, it added a slight wrinkle in that it also provides a Python install.

For some reason, when using the `pyjanitor` package which hooks into `pandas`, it wasn't calling the `pandas` install that's in the virtual env, but instead this one that's loaded by the R-CRAN module instead, which it doesn't work with.

Instead I've gone back to manually creating a library for R packages and instead set the source location when loading the packages with `library`.
This is tedious and manual but it works... Although it hasn't been tested by a user who _doesn't_ have a local R package install yet.