Fixed R dependency issues: Load CRAN module but unload conflicting Python
The R-bundle-CRAN module contains a large number of popular packages from CRAN, easily enough for
this work. However, when it was last used it caused a problem in the time series feature calculation, which makes use of the Python package `pyjanitor` for some data wrangling. This package hooks onto `pandas`, however, rather than accessing the `pandas` installed on the venv, it was trying to access a version that comes with the `SciPy-bundle` module that is a dependency of `R-bundle-CRAN`. This version of pandas is incompatible with `pyjanitor` and raises an error.

Rather than trying to create a local package repository, which was problematic as my user's R environment had been modified externally, it was far simpler to load `R-bundle-CRAN` but unload `SciPy-bundle.`