Ignore errors when creating tiff stack
Tiffs are limited to 4GB, and tiffcp will error (exit status = 1) if the output
stack is larger. This error is now ignored, so that the rest of the pipeline won't fail.
This is done by ignoring all errors raised in this process. `tiffcp` uses an exit status of 1
for all errors, so it isn't easily possible to only ignore this specific 4GB error.