feat: Restructure experiment folder tree (#42)
* Refactor: experiments are sub-divided into sites & image types

Sites are wells, image types are brightfield or phase.
This commit adds several features:

  - Publish to the new directory tree with these subdivisions
  - Run pipeline from experiment directory
  - Have the ability to specify which portions of the pipeline to run

* Updated launcher script & config to handle sites and image types

Also allows for the specification of which parts of the pipeline get run

* Don't update bioldata files modification time

If a Nextflow Pipeline uses cached outputs, such as a mask or CSV file,
it will publish this again to the output directory, updating the
modification time but the contents will be identical. Rclone by default
will update the remote copy with the new modification time. However,
this isn't that useful to us and in fact could be misleading if some
files have a later modification time than when they were actually
generated.

* Store configs in separate 'configs' dir

* Remove the default site name and image type

* Ported QC filtering to R

R has more concise data wrangling functions.
This could be where we remove duplicate TRACK_IDs

* Only display unique TRACK_IDs

* Added R dependencies

* Fixed typo

* Explicitly load venv

* Load reticulate environment before using Python

* Fixed reticulate not loading venv

the specified venv is ignored if an environment variable is set

* Format median values as floats

The median of a set of ints can also be a float, and sprintf will raise an error
if the value is not explicitly typed as a float

* Reduce track_images resources for quicker queueing

This was taking a long time to queue so instead the default resources are
lowered but with an increased amount retry on failure

* Add maven to PATH

* Update README.md

Updated README with description of running the pipeline with the new config setup

* Replaced R CRAN module with custom library

We had a problem in that R doesn't have a neat package manager like pip or
virtual environment manager like venv, yet we needed multiple users to have access to
the exact same R packages when running the pipeline.

Initially I tried to setup a library of packages in the shared project dir, but I was unable to make this the default location as the `R_LIBS_USER` env var that determines where R should find packages
wasn't being read by R - it seemingly only reads it in if it's provided in a .Renviron
file, which resides in a user's home directory.
Rather than requiring every user of the pipeline to install the same R packages, we used
the R-CRAN EasyBuild module, which provides a whole host of packages.
However, while this solved the R problem, it added a slight wrinkle in that it also provides a Python install.

For some reason, when using the `pyjanitor` package which hooks into `pandas`, it wasn't calling the `pandas` install that's in the virtual env, but instead this one that's loaded by the R-CRAN module instead, which it doesn't work with.

Instead I've gone back to manually creating a library for R packages and instead set the source location when loading the packages with `library`.
This is tedious and manual but it works... Although it hasn't been tested by a user who _doesn't_ have a local R package install yet.

* Fixed R dependency issues: Load CRAN module but unload conflicting Python

The R-bundle-CRAN module contains a large number of popular packages from CRAN, easily enough for
this work. However, when it was last used it caused a problem in the time series feature calculation, which makes use of the Python package `pyjanitor` for some data wrangling. This package hooks onto `pandas`, however, rather than accessing the `pandas` installed on the venv, it was trying to access a version that comes with the `SciPy-bundle` module that is a dependency of `R-bundle-CRAN`. This version of pandas is incompatible with `pyjanitor` and raises an error.

Rather than trying to create a local package repository, which was problematic as my user's R environment had been modified externally, it was far simpler to load `R-bundle-CRAN` but unload `SciPy-bundle.`

* Fixed bug in parsing OME tiffs

The output frames weren't zero padded, so the filenames weren't in numerical order (i.e. frame 10 was counted as coming before frame 2)
This has been fixed by simply zero padding the frame numbers output from the OME splitting

* Don't track DS_Store

* Ignore errors when creating tiff stack

Tiffs are limited to 4GB, and tiffcp will error (exit status = 1) if the output
stack is larger. This error is now ignored, so that the rest of the pipeline won't fail.
This is done by ignoring all errors raised in this process. `tiffcp` uses an exit status of 1
for all errors, so it isn't easily possible to only ignore this specific 4GB error.

* Fixed bug with create_tiff_stack not ignoring errors

The configuration associated with the process's label takes precedence
over the configuration provided in the process itself (an interesting design choice),
which means that the 'ignore' error strategy wasn't correctly being applied.
The label has now been split into 2: 'slurm', which uses uses the Slurm executor, and 'slurm_retry' which looks for the Slurm specific error codes to retry on.

The error caused by the maximum tiff file size being reached is now correctly ignored.