feat: Handle tomocube images (#52)
* Remove spaces from filenames

While this could be an issue with any microscope, it hasn't been encountered
until we obtained Tomocube data, which, at least in the configuration we were
using, output filenames with spaces in.

Having spaces in filenames breaks several processes, which rely upon receiving a
single string containing all image filenames delimited by spaces (which is what
is output by the 'collect' channel operator).

The `replace` Unix tool can easily remove these spaces, replacing them with an underscore.

* Use natural sort on filenames

This handles non-zero padded ints (a la Tomocube)

* Removed accidental files

* Only rename files if they contain a space

This resolves the issue of the non-zero exit code when there isn't a space,
as well as the output file from the process being the same as the input.