feat: separate launch dir per timelapse (#53)
* Run each timelapse from its own working directory

This will mean that Nextflow can `-resume` different timelapses
from the same experiment - this wasn't possible before as all timelapses
from an experiment shared a launch dir, which locks the cache db.

This now means that multiple pipelines of the same timelapse can't `-resume`
simultaneously, i.e. if wanting to rerun different segmentation models for
the same dataset.

* Limited rclone output to once a minute updates

Previously it was displaying a message for each file transferred, which
polluted the logs. Now it only displays a progress summary each minute.
This could be disabled entirely and only summarize at the end, but it's
reassuring to have some updates if the transfer is taking a long time.

* Fixed timelapse naming bug

* Nextflow launch dir is timelapse + segmentation + tracking specific