Fix RGB->Grayscale conversion and scale contrast (#59)
There was a bug before in which the `convert('L')` call was attempting
to convert every image into grayscale encoding, regardless of whether
they were RGB or not.
This conversion was added when the ioLight images were added to the
pipeline as they are RGB.
The conversion feature was tested on the test Livecyte image in `CellPhePy`
package and didn't affect it, so it was released.

However, the test Livecyte image had been converted to 8-bit tiff at
some point in time, as they come off the machine at 16-bit.
`convert('L')` doesn't ignore non-RGB images - it still tries to force
a coercion, not just to grayscale but to 8-bit grayscale. This had no
impact on the test Livecyte image, but on the raw images that are encoded
as 16-bit uints, they get truncated at pixel values of 255.
I.e. rather than scaling the image (65,535 -> 255, 32,768 -> 128),
anything higher than 255 in the 16-bit image gets clamped to 255,
resulting in a huge loss of information.

This issue is fixed by checking for RGB before casting to grayscale.

In addition, this patch also changes the normalization behaviour.
Previously, images were scaled to [0,1] by dividing by 255, again
because we were using 8-bit images.
This behaviour has been changed to shift the contrast, so that 1
now indicates the maximum intensity in the image, and 0 means the
minimum.
This is because several microscopes output images scaled in this way
already (ioLight & Livecyte) but the Tomocube doesn't. Instead, the
Tomocube pixel intensities were in a very small range. This change
brings the Tomocube in line with the other microscopes.

This change is going to be added upstream to the CellPhePy package soon,
but it will take longer to get released there so it is being pushed out
on the pipeline first to allow us to re-process the Livecyte datasets.