First effort in obtaining unique TRACK_IDs per frame
A graph is formed of all the nodes, which is traversed recurively to
build up a list of tracks, where each track only has 1 cell id per
frame associated with it.
This is the default PhaseFocus behaviour, but isn't the case in
TrackMate, where multiple cells in the same frame were assigned the same
track id, as Trackmate treats track id as an identifier of lineage.

NB: this is a WIP, and doesn't fully work. It is currently outputting
multiple TRACK_IDs for the same cell in the same frame.