Replaced segmentation highlighting with borders
Previously the cells were indicated by having a transparent colour fill
overlaid onto the original image. This made it hard to see what the cell
looked like in the raw image, as the view was obscured through the
colour fill. This change replaces that with a yellow line of the cell
border.