feat: Replaced segmentation highlighting with borders (#65)
* Replaced segmentation highlighting with borders

Previously the cells were indicated by having a transparent colour fill
overlaid onto the original image. This made it hard to see what the cell
looked like in the raw image, as the view was obscured through the
colour fill. This change replaces that with a yellow line of the cell
border.

* Handle relative path to config

* Made cell boundaries more visible by enlargening

* Can select segmentation highlighting method in config

Can either use:

  - 'outline': Draws the cell border in yellow. Default.
  - 'fill': Fills cells with a solid outline. The old behaviour

* Defaults to 'outline' segmentation overlay when not correctly supplied

* Added segmentation_highlight option to templates