Only allow ROIs that have 3 pixels in both dims
Ideally this check would be done in the existing `filter_size_and_observations`
process, but that only has access to the Trackmate features which don't contain
dimension sizes. Rather than passing in the ROIs to `filter_size_and_observation` and making it read through every ROI file, it's more straight forward to handle it in `frame_features`.