feat: additional resources tomocube (#66)
* Assign more resources to Tomocube runs

The Tomocube requires far more resources, particularly on Frame Feature extraction, than the LiveCyte or ioLight. This is likely a direct result of the cell crops being far larger owing to the increased magnification.

* Fixed syntax

* Only allow ROIs that have 3 pixels in both dims

Ideally this check would be done in the existing `filter_size_and_observations`
process, but that only has access to the Trackmate features which don't contain
dimension sizes. Rather than passing in the ROIs to `filter_size_and_observation` and making it read through every ROI file, it's more straight forward to handle it in `frame_features`.