hotfix: don't calculate Trackmate spot ellipse features (#67)
* Don't error if masks dir already exists

* Removed ellipse features

There was a bug that affected a small number of datasets, whereby
the tracking would hang on the spot feature calculation stage.
This was identified to be due to some ellipse features.
Previously, all feature calculators were added with
`addAllAnalyzers`, but it's not possible to just remove the factory
that provides the ellipse features.
Instead, we'll manually add all feature factories except the ellipse
one.