feat: Longship migration (#68)
* Moving raw data to Longship, outputs to scratch

* Move data to longship mounted on research0

* fixes to paths

* WIP: Nextflow can see the images from longship

But Python can't read them (PIL error when reading file).
This shouldn't be due to the 'file' being a symlink as this is how
Nextflow always stores files and how it was previously. It might be a
permissions issue, as the longship group isn't named

* Renamed variable to be more descriptive

* Copy data back to bioldata via longship

Rather than copying data directly from Scratch -> bioldata, instead the
pipeline outputs are moved from Scratch to Longship on the Viking end,
then copied from Longship to bioldata on research0.
This has several advantages:

  - Should be quicker as Longship is mounted on both Viking and Research0
  - Keeps Scratch clean, only containing the Nextflow cache
  - Keeps a duplicate copy on Longship for quicker processing when needed

* Pipeline raw data dir and output dir are now params

Previously they were derived at runtime from the path of the config.
This was very restrictive and meant that the workflow only works within our
folder structure of config file etc...
It is now far more flexible.

* Fixed QC paths to longship

* Always attempt to resume a pipeline

* Removed debug prints

* Add template for Tomocube Brightfield

Just uses the Livecyte BF model with an increased diameter