Run pipeline locally, not via Slurm
This is partly for testing purposes, as the HPC is very busy at the moment,
but also because nothing of computational significance runs in the main workflow,
so it might as well run on the login node. This removes the need to estimate
how long the full pipeline will take.