feat: Use Apptainer containers to run Nextflow processes (#70)
* Ensure clean env

* Add option to process segmentation in batches on a GPU
rather than parallelised over CPU cores as is the default

* Always request max cores & memory on GPU queue

On GPU nodes you are limited to 32 cores and 176GB RAM per GPU, hence
there is no reason to not request less than this.

* If need more than 30mins, use main GPU queue

gpu_short is limited to 30 mins

* Raw data directory is always relative to config directory

Previously the raw data directory was hardcoded to be on Longship, even
if the config was on scratch. This is now more flexible as it "only" requires
that the experiment structure is <Experiment>/configs and <Experiments>/raw,
regardless of which filesystem they're on

* Added template for SAM model that runs on GPU

* Use GPU queue for jobs > 30 mins

* Working containers for segmentation

* Remove hardcoding of CPUs - by default GPUs now use all available

* Run `jq` processes in container

NB: rather than using the [latest images from the jq dev team](https://github.com/jqlang/jq/pkgs/container/jq),
we'll use the only image from an early version uploaded 10 years ago
by the [initial developer](https://hub.docker.com/r/stedolan/jq/tags).
This isn't ideal, but the latest images are stripped back and only
contain the executable, but we also need `bash` as Nextflow requires this
to launch commands in containers.

Usin this old version of `jq` also has a mandatory positional argument for
the filter, hence adding the `'.'` to the call

* Run pipeline locally, not via Slurm

This is partly for testing purposes, as the HPC is very busy at the moment,
but also because nothing of computational significance runs in the main workflow,
so it might as well run on the login node. This removes the need to estimate
how long the full pipeline will take.

* Use container for `awk` process

Using a Debian slim image as this comes with `awk` as well as `/bin/bash`
at the required path.

* Removed cellphe as a dependency

Instead directly load the `roifile` package

* Removed cellphe as a dependency

Export the relevant code snippets from the CellPhe python package so
that the only dependencies are `scyjava` and `pyimagej`

* WIP: Run QC reports in Quarto container

Not fully working as the rendering complains about SVG not being available
but a) this shouldn't matter as we're building HTML and b) it worked fine
locally

* Use venv from container

* Remove Viking-specific modules

These dependencies are now bundled through containers instead

* Update QC reports to work with Alpine images

These don't use a virtual env, don't have tidyverse installed, and don't
have the default cairo that comes bundled with R.

* Working tracking

* Working parse_xml

* Added container for every process (untested)

* Fixed frame-level features

* Fixed time-series feats & combining frame feats

* Fix container

* Confirmed containers have access to all CPUs in cgroup

* Removed all Viking module deps except NextFlow and Apptainer

The pipeline now only has these 2 dependencies, making it much
more portable. The remaining task to make it fully portable will be
to parameterize away all the Viking-specific config

* Write quarto cache to container rather than mounted volume

* Use cellpose's `gpu` flag to determine if to run on gpu or not

Previously we had created our own flag, but this was redundant as
cellpose already has a flag so we can just use this one.