Use updated cellpose images that contain `ps`
Nextflow requires that containers can run `ps`, which isn't included
in Debian slim images, which the cellpose ones are based on. The
[cellpose images were updated to 0.1.2 to include ps](https://github.com/uoy-research/cellphe-pipeline-images/pull/19),
so the templates need to point at these latest versions too.

Interestingly, when running Nextflow locally it didn't mind that ps
wasn't available, it was only on Viking it mattered.