feat: multiple configuration profiles (#73)
* Moving Slurm specific config to viking profile

This `york_viking` profile contains all Slurm/Viking specific setup,
including queue names, retry strategy, number of cpu resources etc...

* Reordered process directives

Standardizd process directives ordering:

  - container
  - containeroptions (if used)
  - publishdir (if used)
  - label (if used)
  - time
  - memory

* Removed loading venv and setting maven path

Both Python environment and maven should now be included within
containers so they don't need to be setup for the whole pipeline.

* Update process_dataset.sh to point to renamed pipeline file

Ideally this script would either live in a separate repo, allowing this
repo to just contain the pipeline and a separate (private) repo just
contains the UoY specific setup - or it lives in a sub-dir called
`scripts` or something. Basically want to signpost directly to the
pipeline itself, rather than the UoY specific scripts around it

* Use custom jq image

The previous image used manifest V1 and isn't supported by apptainer

* Bumped to latest images

* Fixed syntax error in container definition

It meant that the container was being ignored

* Moved all resource requirements to Viking profile

I was under the impression from the docs that resource requirements
would be ignored when run locally. This is not the case and Nextflow
will error if you ask for e.g. 64GB RAM and there is only 16GB
available.

* Use relative path for Quarto file paths

Can't use shebang with Quarto as the --- XXXX --- front matter *must*
be the first contents in the file, hence this hacky work around.

* Handle non-numeric requested meomry

E.g. when running locally and there is no 'task.memory' value, so 'null'
gets passed through to the Python script.

* Added notes to docs to expand upon later

* Corrected env var for apptainer temp dir

* Use Viking profile when running pipeline on Viking

* Fixed config syntax

key and values must be separated with an `=` in the .config file, but
this isn't the case when the same options are provided in the process
definition.

* Adds ability to specify the Cellpose version by means of docker image

This means that users can now easily switch between cellpose 3 and 4,
running either on CPU or GPU as required.

* Replaced site and image_type with timelapse_id

* Fixed nextflow syntax error

* Templates use our cellpose images and internal custom models

* Moved UoY specific scripts to `scripts` sub-folder

* Point Numba cache at /tmp

By default Numba tries to cache in the package installation dir, which
is in /usr/local/lib, which isn't writable inside apptainer.
See https://github.com/numba/numba/issues/9312

* Use updated cellpose images that contain `ps`

Nextflow requires that containers can run `ps`, which isn't included
in Debian slim images, which the cellpose ones are based on. The
[cellpose images were updated to 0.1.2 to include ps](https://github.com/uoy-research/cellphe-pipeline-images/pull/19),
so the templates need to point at these latest versions too.

Interestingly, when running Nextflow locally it didn't mind that ps
wasn't available, it was only on Viking it mattered.

* Also fix numba cache dir for segment_gpu process

* Use project-wide cellpose cache on UoY HPC

By default cellpose tries to save cache in the user's home directory,
which isn't mounted in Nextflow. This allows the use of a (writable) cache
in the project dir, which also saves having to download the models each time
for a new user.

* Fixed segmentation config on Viking profile

The `containerOptions` in the Viking profile were *overwriting* the ones
present in `main.nf`, not *appending* to them as I had assumed would
happen.

* Fixed = syntax issue in config

* First pass at README

* Second pass on README

* README edit

* Added flowchart diagram to README

* Using mermaid version of flowchart

* Using ome.companion dataset as input for flowchart

* README polish

* Fixing section links

* Further README polish